success parser test

+++ src/main/java/telegrambot/model/User.kt
@@ -1,6 +1,6 @@
 package telegrambot.model
 
-data class User(val id : Int,
+data class User(val id : Long,
                 val isBot : Boolean,
                 val firstName : String,
                 val secondName : String,
